emit abstract method bodies with a check for previous definition (#255)
* emit abstract method bodies with a check for previous definition

* add support for abstract static methods

* add test for abstract methods that does not work yet

* add TODO markers for abstract method definedness checks

* explicitly mark a file as compiled: false

* add expectation files for abstract methods

+++ test/testdata/compiler/impl_abstract_via_include.rb
@@ -7,7 +7,7 @@ module IFoo
   extend T::Helpers
   interface!
 
-  sig {abstract.returns(Integer)}
+  sig {abstract.returns(Float)}
   def foo; end
 end
